+++ HISTORY.md
@@ -1,5 +1,13 @@
 # History of versions #
 
+* Version 1.20.0 (2022-11-20)
+  * Fixed Guake window zero height on first run
+  * Fixed Guake keybinding failing until re-set
+  * Added hotkeys to close a tile and a tab
+  * Now asks for confirmation when trying to close a terminal with a running program
+  * Fixed color scheme selection
+  * Allows to change the top bar colors
+  * Cleaned the color management code
 * Version 1.19.1 (2022-10-13)
   * Added support for Gnome Shell 43
 * Version 1.19.0 (2022-09-28)